==> code/03_summarize_pilotsample.ipynb <==
library(dplyr)

# 1. Read in files for each of the DCPS schools and create strata

## read in
teachers_chechs = read.csv("../randomization_status/teachers_in_pilot_FORLAB.csv")
teachers_chechs

n_strata = length(unique(teachers_chechs$pair_id_SG))
teachers_chechs = teachers_chechs %>%
                mutate(strata_id = sprintf("checHS_strata_%s",
                                          rep(seq(1, n_strata), each = 2)))

teachers_chechs_tomerge = teachers_chechs %>%
                    dplyr::select(Teacher.Email, treatment_status_SG, strata_id) %>%
                    rename(teacher_id = Teacher.Email, 
                          treatment = treatment_status_SG)



## summarise
table(teachers_chechs$treatment_status_SG)
teachers_chechs %>% group_by(treatment_status_SG) %>% summarise(n = sum(total_students))

teachers_checms = read.csv("../randomization_status/chec_MS_teacherpilotstatus.csv")


n_strata = length(unique(teachers_checms$strata_name))
teachers_checms_tomerge = teachers_checms %>%
                    mutate(strata_id = sprintf("checMS_strata_%s",
                                          as.numeric(factor(strata_name)))) %>%
                    dplyr::select(teacher_id, treatment, strata_id) 



teachers_anacostia = read.csv("../randomization_status/anacostia_teacherpilotstatus.csv")


head(teachers_anacostia)

teachers_anacostia %>% group_by(treatment) %>% summarise(n = sum(total_students))

teachers_ana_tomerge = teachers_anacostia %>%
                    mutate(strata_id = sprintf("anacostia_strata_%s",
                                          as.numeric(factor(strata_name)))) %>%
                    dplyr::select(teacher_id, treatment, strata_id) 



teachers_dunbar = read.csv("../randomization_status/dunbar_teacherpilotstatus.csv")



teachers_dunbar_tomerge = teachers_dunbar %>%
                    mutate(strata_id = sprintf("dunbar_strata_%s",
                                          as.numeric(factor(strata_name)))) %>%
                    dplyr::select(teacher_id, treatment, strata_id) 

teachers_dunbar %>% group_by(treatment) %>% summarise(n = sum(total_students))
head(teachers_dunbar)

# 2. Create one master file with randomization status of all teachers

files_torbind =  ls(pattern = "tomerge")
objects_torbind = lapply(files_torbind, function(x) get(x))

teachers_fallsem = do.call(rbind.data.frame, objects_torbind)
write.csv(teachers_fallsem, "../randomization_status/fallsemschools_summarizeRA.csv",
         row.names = FALSE)
